consumer/requirements.txt
fastapi==0.104.1
kafka_python==2.0.2
uvicorn